requirements.txt
absl-py==1.3.0
affine==2.3.1
astroid==2.12.13
astunparse==1.6.3
attrs==22.1.0
cachetools==5.2.0
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.1.3
click-plugins==1.1.1
cligj==0.7.2
dill==0.3.6
flatbuffers==22.12.6
gast==0.4.0
google-auth==2.15.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.51.1
h5py==3.7.0
idna==3.4
imageio==2.23.0
isort==5.11.3
joblib==1.2.0
keras==2.11.0
lazy-object-proxy==1.8.0
libclang==14.0.6
Markdown==3.4.1
MarkupSafe==2.1.1
mccabe==0.7.0
networkx==2.8.8
numpy==1.24.0
oauthlib==3.2.2
opt-einsum==3.3.0
packaging==22.0
Pillow==9.3.0
platformdirs==2.6.0
protobuf==3.19.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pylint==2.15.9
pyparsing==3.0.9
PyWavelets==1.4.1
rasterio==1.3.4
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
scikit-image==0.19.3
scikit-learn==1.2.0
scipy==1.9.3
six==1.16.0
snuggs==1.4.7
tensorboard==2.11.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.11.0
tensorflow-estimator==2.11.0
tensorflow-io-gcs-filesystem==0.29.0
termcolor==2.1.1
threadpoolctl==3.1.0
tifffile==2022.10.10
tomli==2.0.1
tomlkit==0.11.6
typing_extensions==4.4.0
urllib3==1.26.13
Werkzeug==2.2.2
wrapt==1.14.1
